+++ R/reverse_species_order.R
@@ -15,7 +15,7 @@
 #' 
 #' 
 #' @examples 
-#' # basic usage of load_orthologs :
+#' # basic usage of reverse_species_order :
 #' 
 #' orthologs_file <- system.file("extdata","Bflo_vs_Pech.tab",package="macrosyntR")
 #' bedfile_sp1 <- system.file("extdata","Bflo.protein_products.bed",package="macrosyntR")